refactor(core): Split validActions/cards.ts into focused modules for LLM optimization
Split the 746-line cards.ts file into a modular directory structure:
- cards/index.ts: Re-exports and getCard() function
- cards/combat.ts: getPlayableCardsForCombat and combat phase playability
- cards/normalTurn.ts: getPlayableCardsForNormalTurn and normal turn playability
- cards/manaPayment.ts: canPayForSpellBasic, findPayableManaColor
- cards/effectDetection/: Effect detection functions split into:
  - combatEffects.ts: effectHasAttack, effectHasBlock, effectHasRangedOrSiege
  - movementEffects.ts: effectHasMove, effectHasInfluence
  - resourceEffects.ts: effectHasHeal, effectHasDraw, effectHasModifier, effectHasManaGain
  - specialEffects.ts: effectHasManaDrawPowered, effectHasCrystal, effectHasCardBoost, effectHasEnemyTargeting
  - index.ts: Re-exports all effect detection functions and effectIsUtility

All modules are under 200 lines. No behavior changes.
