--- packages/core/src/engine/validActions/cards.ts
@@ -1,746 +1,0 @@
-/**
- * Card playability computation for combat
- *
- * Determines which cards in the player's hand can be played during each combat phase.
- */
-
-import type { GameState } from "../../state/GameState.js";
-import type { Player } from "../../types/player.js";
-import type { CombatState, CombatPhase } from "../../types/combat.js";
-import type { CardEffect, DeedCard } from "../../types/cards.js";
-import type { PlayCardOptions, PlayableCard, ManaColor, SidewaysOption } from "@mage-knight/shared";
-import {
-  COMBAT_TYPE_RANGED,
-  COMBAT_TYPE_SIEGE,
-  PLAY_SIDEWAYS_AS_ATTACK,
-  PLAY_SIDEWAYS_AS_BLOCK,
-} from "@mage-knight/shared";
-import { canPayForMana, canPayForTwoMana } from "./mana.js";
-import {
-  EFFECT_GAIN_ATTACK,
-  EFFECT_GAIN_BLOCK,
-  EFFECT_CHOICE,
-  EFFECT_COMPOUND,
-  EFFECT_CONDITIONAL,
-  EFFECT_SCALING,
-  EFFECT_SELECT_COMBAT_ENEMY,
-} from "../../types/effectTypes.js";
-import {
-  COMBAT_PHASE_RANGED_SIEGE,
-  COMBAT_PHASE_BLOCK,
-  COMBAT_PHASE_ATTACK,
-} from "../../types/combat.js";
-import { getBasicActionCard } from "../../data/basicActions/index.js";
-import { getAdvancedActionCard } from "../../data/advancedActions/index.js";
-import { getSpellCard } from "../../data/spells.js";
-import { DEED_CARD_TYPE_WOUND, DEED_CARD_TYPE_SPELL } from "../../types/cards.js";
-import { describeEffect } from "../effects/describeEffect.js";
-
-import { MANA_BLACK } from "@mage-knight/shared";
-
-/**
- * Check if a player can pay for a spell's basic effect.
- *
- * Spells require mana of their color even for the basic effect (unlike action cards).
- * The spell's color is found in poweredBy (the non-black color).
- */
-function canPayForSpellBasic(
-  state: GameState,
-  player: Player,
-  card: DeedCard
-): boolean {
-  // Find the spell's color (the non-black color in poweredBy)
-  const spellColor = card.poweredBy.find((c) => c !== MANA_BLACK);
-  if (!spellColor) return false;
-
-  return canPayForMana(state, player, spellColor);
-}
-
-/**
- * Check if a player can power a card.
- *
- * For action cards: returns the first mana color from poweredBy that the player can pay for.
- * For spells: returns the spell's color ONLY if the player can pay for BOTH black AND that color.
- *
- * Returns undefined if the card cannot be powered.
- */
-function findPayableManaColor(
-  state: GameState,
-  player: Player,
-  card: DeedCard
-): ManaColor | undefined {
-  if (card.poweredBy.length === 0) return undefined;
-
-  // Spells require BOTH black mana AND the spell's color
-  if (card.cardType === DEED_CARD_TYPE_SPELL) {
-    // Spell's poweredBy should be [MANA_BLACK, spell_color]
-    const spellColor = card.poweredBy.find((c) => c !== MANA_BLACK);
-    if (!spellColor) return undefined;
-
-    // Check if player can pay for both black AND the spell color
-    if (canPayForTwoMana(state, player, MANA_BLACK, spellColor)) {
-      return spellColor; // Return the spell color (black is implicit)
-    }
-    return undefined;
-  }
-
-  // Action cards: any one of the poweredBy colors works (OR logic)
-  return card.poweredBy.find((color) => canPayForMana(state, player, color));
-}
-
-/**
- * Get playable cards for combat based on the current phase.
- */
-export function getPlayableCardsForCombat(
-  state: GameState,
-  player: Player,
-  combat: CombatState
-): PlayCardOptions {
-  const cards: PlayableCard[] = [];
-
-  for (const cardId of player.hand) {
-    const card = getCard(cardId);
-    if (!card) continue;
-
-    // Wounds cannot be played
-    if (card.cardType === DEED_CARD_TYPE_WOUND) continue;
-
-    const playability = getCardPlayabilityForPhase(card, combat.phase);
-
-    // Check resolvability - effect must actually be able to do something
-    const basicIsResolvable = isEffectResolvable(state, player.id, card.basicEffect);
-    const poweredIsResolvable = isEffectResolvable(state, player.id, card.poweredEffect);
-
-    // For spells, basic effect also requires mana (the spell's color)
-    // Get the spell's color from poweredBy (excluding black)
-    const spellBasicManaAvailable = card.cardType === DEED_CARD_TYPE_SPELL
-      ? canPayForSpellBasic(state, player, card)
-      : true; // Action cards don't need mana for basic effect
-
-    // Can only play basic if the phase allows it AND the effect is resolvable
-    // AND for spells, the player has the spell's color mana
-    const canActuallyPlayBasic = playability.canPlayBasic && basicIsResolvable && spellBasicManaAvailable;
-
-    // Check if the card has a powered effect for this phase AND player can pay for it AND it's resolvable
-    const payableManaColor = (playability.canPlayPowered && poweredIsResolvable)
-      ? findPayableManaColor(state, player, card)
-      : undefined;
-    const canActuallyPlayPowered = payableManaColor !== undefined;
-
-    if (canActuallyPlayBasic || canActuallyPlayPowered || playability.canPlaySideways) {
-      const playableCard: PlayableCard = {
-        cardId,
-        name: card.name,
-        canPlayBasic: canActuallyPlayBasic,
-        canPlayPowered: canActuallyPlayPowered,
-        canPlaySideways: playability.canPlaySideways,
-        basicEffectDescription: describeEffect(card.basicEffect),
-        poweredEffectDescription: describeEffect(card.poweredEffect),
-      };
-
-      // Only add optional properties when they have values
-      if (payableManaColor && canActuallyPlayPowered) {
-        (playableCard as { requiredMana?: ManaColor }).requiredMana = payableManaColor;
-      }
-      if (card.cardType === DEED_CARD_TYPE_SPELL) {
-        (playableCard as { isSpell?: boolean }).isSpell = true;
-      }
-      if (playability.sidewaysOptions && playability.sidewaysOptions.length > 0) {
-        (playableCard as { sidewaysOptions?: readonly SidewaysOption[] }).sidewaysOptions = playability.sidewaysOptions;
-      }
-
-      cards.push(playableCard);
-    }
-  }
-
-  return { cards };
-}
-
-interface CardPlayability {
-  canPlayBasic: boolean;
-  canPlayPowered: boolean;
-  canPlaySideways: boolean;
-  sidewaysOptions: SidewaysOption[];
-}
-
-/**
- * Check if an effect is a "utility" effect that can be played during any combat phase.
- * These include mana generation, card draws, healing, card boost, and enemy-targeting spells.
- */
-function effectIsUtility(effect: CardEffect): boolean {
-  return (
-    effectHasManaGain(effect) ||
-    effectHasDraw(effect) ||
-    effectHasHeal(effect) ||
-    effectHasModifier(effect) ||
-    effectHasManaDrawPowered(effect) ||
-    effectHasCardBoost(effect) ||
-    effectHasCrystal(effect) ||
-    effectHasEnemyTargeting(effect)
-  );
-}
-
-/**
- * Determine if a card can be played in a specific combat phase.
- */
-function getCardPlayabilityForPhase(
-  card: DeedCard,
-  phase: CombatPhase
-): CardPlayability {
-  switch (phase) {
-    case COMBAT_PHASE_RANGED_SIEGE:
-      return {
-        // Ranged/siege phase: can play for ranged/siege attack OR utility effects
-        canPlayBasic: effectHasRangedOrSiege(card.basicEffect) || effectIsUtility(card.basicEffect),
-        canPlayPowered: effectHasRangedOrSiege(card.poweredEffect) || effectIsUtility(card.poweredEffect),
-        canPlaySideways: false, // Can't play sideways for ranged/siege
-        sidewaysOptions: [],
-      };
-
-    case COMBAT_PHASE_BLOCK:
-      return {
-        // Block phase: can play for block OR utility effects
-        canPlayBasic: effectHasBlock(card.basicEffect) || effectIsUtility(card.basicEffect),
-        canPlayPowered: effectHasBlock(card.poweredEffect) || effectIsUtility(card.poweredEffect),
-        canPlaySideways: card.sidewaysValue > 0,
-        sidewaysOptions: card.sidewaysValue > 0
-          ? [{ as: PLAY_SIDEWAYS_AS_BLOCK, value: card.sidewaysValue }]
-          : [],
-      };
-
-    case COMBAT_PHASE_ATTACK:
-      return {
-        // Attack phase: can play for attack OR utility effects
-        canPlayBasic: effectHasAttack(card.basicEffect) || effectIsUtility(card.basicEffect),
-        canPlayPowered: effectHasAttack(card.poweredEffect) || effectIsUtility(card.poweredEffect),
-        canPlaySideways: card.sidewaysValue > 0,
-        sidewaysOptions: card.sidewaysValue > 0
-          ? [{ as: PLAY_SIDEWAYS_AS_ATTACK, value: card.sidewaysValue }]
-          : [],
-      };
-
-    default:
-      // ASSIGN_DAMAGE phase - no cards played
-      return {
-        canPlayBasic: false,
-        canPlayPowered: false,
-        canPlaySideways: false,
-        sidewaysOptions: [],
-      };
-  }
-}
-
-/**
- * Check if an effect provides ranged or siege attack.
- */
-function effectHasRangedOrSiege(effect: CardEffect): boolean {
-  switch (effect.type) {
-    case EFFECT_GAIN_ATTACK:
-      return effect.combatType === COMBAT_TYPE_RANGED || effect.combatType === COMBAT_TYPE_SIEGE;
-
-    case EFFECT_CHOICE:
-      return effect.options.some(opt => effectHasRangedOrSiege(opt));
-
-    case EFFECT_COMPOUND:
-      return effect.effects.some(eff => effectHasRangedOrSiege(eff));
-
-    case EFFECT_CONDITIONAL:
-      return effectHasRangedOrSiege(effect.thenEffect) ||
-        (effect.elseEffect ? effectHasRangedOrSiege(effect.elseEffect) : false);
-
-    case EFFECT_SCALING:
-      return effectHasRangedOrSiege(effect.baseEffect);
-
-    default:
-      return false;
-  }
-}
-
-/**
- * Check if an effect provides block.
- */
-function effectHasBlock(effect: CardEffect): boolean {
-  switch (effect.type) {
-    case EFFECT_GAIN_BLOCK:
-      return true;
-
-    case EFFECT_CHOICE:
-      return effect.options.some(opt => effectHasBlock(opt));
-
-    case EFFECT_COMPOUND:
-      return effect.effects.some(eff => effectHasBlock(eff));
-
-    case EFFECT_CONDITIONAL:
-      return effectHasBlock(effect.thenEffect) ||
-        (effect.elseEffect ? effectHasBlock(effect.elseEffect) : false);
-
-    case EFFECT_SCALING:
-      return effectHasBlock(effect.baseEffect);
-
-    default:
-      return false;
-  }
-}
-
-/**
- * Check if an effect provides any attack (melee, ranged, or siege).
- */
-function effectHasAttack(effect: CardEffect): boolean {
-  switch (effect.type) {
-    case EFFECT_GAIN_ATTACK:
-      return true;
-
-    case EFFECT_CHOICE:
-      return effect.options.some(opt => effectHasAttack(opt));
-
-    case EFFECT_COMPOUND:
-      return effect.effects.some(eff => effectHasAttack(eff));
-
-    case EFFECT_CONDITIONAL:
-      return effectHasAttack(effect.thenEffect) ||
-        (effect.elseEffect ? effectHasAttack(effect.elseEffect) : false);
-
-    case EFFECT_SCALING:
-      return effectHasAttack(effect.baseEffect);
-
-    default:
-      return false;
-  }
-}
-
-/**
- * Get a card definition by ID.
- * Supports basic action, advanced action, and spell cards.
- */
-export function getCard(cardId: string): DeedCard | null {
-  // Try basic action cards first
-  try {
-    return getBasicActionCard(cardId as Parameters<typeof getBasicActionCard>[0]);
-  } catch {
-    // Not a basic action, continue
-  }
-
-  // Try advanced action cards
-  try {
-    return getAdvancedActionCard(cardId as Parameters<typeof getAdvancedActionCard>[0]);
-  } catch {
-    // Not an advanced action either
-  }
-
-  // Try spell cards
-  const spell = getSpellCard(cardId as Parameters<typeof getSpellCard>[0]);
-  if (spell) {
-    return spell;
-  }
-
-  // Card not found - might be artifact, etc.
-  // TODO: Add support for other card types
-  return null;
-}
-
-// ============================================================================
-// Normal Turn Card Playability (non-combat)
-// ============================================================================
-
-import {
-  PLAY_SIDEWAYS_AS_MOVE,
-  PLAY_SIDEWAYS_AS_INFLUENCE,
-} from "@mage-knight/shared";
-import {
-  EFFECT_GAIN_MOVE,
-  EFFECT_GAIN_INFLUENCE,
-  EFFECT_GAIN_HEALING,
-  EFFECT_GAIN_MANA,
-  EFFECT_DRAW_CARDS,
-  EFFECT_APPLY_MODIFIER,
-  EFFECT_MANA_DRAW_POWERED,
-  EFFECT_GAIN_CRYSTAL,
-  EFFECT_CONVERT_MANA_TO_CRYSTAL,
-  EFFECT_CARD_BOOST,
-} from "../../types/effectTypes.js";
-import { isEffectResolvable } from "../effects/index.js";
-
-/**
- * Get playable cards for normal (non-combat) turns.
- *
- * During a normal turn, cards can provide:
- * - Move points
- * - Influence points
- * - Healing
- * - Sideways: +1 Move or +1 Influence
- */
-export function getPlayableCardsForNormalTurn(
-  state: GameState,
-  player: Player
-): PlayCardOptions {
-  const cards: PlayableCard[] = [];
-
-  for (const cardId of player.hand) {
-    const card = getCard(cardId);
-    if (!card) continue;
-
-    // Wounds cannot be played
-    if (card.cardType === DEED_CARD_TYPE_WOUND) continue;
-
-    const playability = getCardPlayabilityForNormalTurn(state, player.id, card);
-
-    // For spells, basic effect also requires mana (the spell's color)
-    const spellBasicManaAvailable = card.cardType === DEED_CARD_TYPE_SPELL
-      ? canPayForSpellBasic(state, player, card)
-      : true;
-    const canActuallyPlayBasic = playability.canPlayBasic && spellBasicManaAvailable;
-
-    // Check if the card has a powered effect AND player can pay for it
-    const payableManaColor = playability.canPlayPowered
-      ? findPayableManaColor(state, player, card)
-      : undefined;
-    const canActuallyPlayPowered = payableManaColor !== undefined;
-
-    if (canActuallyPlayBasic || canActuallyPlayPowered || playability.canPlaySideways) {
-      const playableCard: PlayableCard = {
-        cardId,
-        name: card.name,
-        canPlayBasic: canActuallyPlayBasic,
-        canPlayPowered: canActuallyPlayPowered,
-        canPlaySideways: playability.canPlaySideways,
-        basicEffectDescription: describeEffect(card.basicEffect),
-        poweredEffectDescription: describeEffect(card.poweredEffect),
-      };
-
-      // Only add optional properties when they have values
-      if (payableManaColor && canActuallyPlayPowered) {
-        (playableCard as { requiredMana?: ManaColor }).requiredMana = payableManaColor;
-      }
-      if (card.cardType === DEED_CARD_TYPE_SPELL) {
-        (playableCard as { isSpell?: boolean }).isSpell = true;
-      }
-      if (playability.sidewaysOptions && playability.sidewaysOptions.length > 0) {
-        (playableCard as { sidewaysOptions?: readonly SidewaysOption[] }).sidewaysOptions = playability.sidewaysOptions;
-      }
-
-      cards.push(playableCard);
-    }
-  }
-
-  return { cards };
-}
-
-/**
- * Determine if a card can be played during normal (non-combat) turn.
- *
- * An effect is playable if:
- * 1. It has a useful effect type (move, influence, heal, draw)
- * 2. The effect is actually resolvable given current game state
- *    (e.g., draw requires cards in deck, heal requires wounds)
- */
-function getCardPlayabilityForNormalTurn(
-  state: GameState,
-  playerId: string,
-  card: DeedCard
-): CardPlayability {
-  // Check if basic effect has a useful effect type AND is resolvable
-  const basicHasUsefulEffect =
-    effectHasMove(card.basicEffect) ||
-    effectHasInfluence(card.basicEffect) ||
-    effectHasHeal(card.basicEffect) ||
-    effectHasDraw(card.basicEffect) ||
-    effectHasManaGain(card.basicEffect) ||
-    effectHasModifier(card.basicEffect) ||
-    effectHasCrystal(card.basicEffect);
-
-  const basicIsResolvable = isEffectResolvable(state, playerId, card.basicEffect);
-
-  // Check if powered effect has a useful effect type AND is resolvable
-  const poweredHasUsefulEffect =
-    effectHasMove(card.poweredEffect) ||
-    effectHasInfluence(card.poweredEffect) ||
-    effectHasHeal(card.poweredEffect) ||
-    effectHasDraw(card.poweredEffect) ||
-    effectHasManaGain(card.poweredEffect) ||
-    effectHasModifier(card.poweredEffect) ||
-    effectHasManaDrawPowered(card.poweredEffect) ||
-    effectHasCrystal(card.poweredEffect) ||
-    effectHasCardBoost(card.poweredEffect);
-
-  const poweredIsResolvable = isEffectResolvable(state, playerId, card.poweredEffect);
-
-  // Sideways options for normal turn: move or influence (always available)
-  const sidewaysOptions: SidewaysOption[] = [];
-  if (card.sidewaysValue > 0) {
-    sidewaysOptions.push({ as: PLAY_SIDEWAYS_AS_MOVE, value: card.sidewaysValue });
-    sidewaysOptions.push({ as: PLAY_SIDEWAYS_AS_INFLUENCE, value: card.sidewaysValue });
-  }
-
-  return {
-    canPlayBasic: basicHasUsefulEffect && basicIsResolvable,
-    canPlayPowered: poweredHasUsefulEffect && poweredIsResolvable,
-    canPlaySideways: card.sidewaysValue > 0,
-    sidewaysOptions,
-  };
-}
-
-/**
- * Check if an effect provides move points.
- */
-function effectHasMove(effect: CardEffect): boolean {
-  switch (effect.type) {
-    case EFFECT_GAIN_MOVE:
-      return true;
-
-    case EFFECT_CHOICE:
-      return effect.options.some(opt => effectHasMove(opt));
-
-    case EFFECT_COMPOUND:
-      return effect.effects.some(eff => effectHasMove(eff));
-
-    case EFFECT_CONDITIONAL:
-      return effectHasMove(effect.thenEffect) ||
-        (effect.elseEffect ? effectHasMove(effect.elseEffect) : false);
-
-    case EFFECT_SCALING:
-      return effectHasMove(effect.baseEffect);
-
-    default:
-      return false;
-  }
-}
-
-/**
- * Check if an effect provides influence points.
- */
-function effectHasInfluence(effect: CardEffect): boolean {
-  switch (effect.type) {
-    case EFFECT_GAIN_INFLUENCE:
-      return true;
-
-    case EFFECT_CHOICE:
-      return effect.options.some(opt => effectHasInfluence(opt));
-
-    case EFFECT_COMPOUND:
-      return effect.effects.some(eff => effectHasInfluence(eff));
-
-    case EFFECT_CONDITIONAL:
-      return effectHasInfluence(effect.thenEffect) ||
-        (effect.elseEffect ? effectHasInfluence(effect.elseEffect) : false);
-
-    case EFFECT_SCALING:
-      return effectHasInfluence(effect.baseEffect);
-
-    default:
-      return false;
-  }
-}
-
-/**
- * Check if an effect provides healing.
- */
-function effectHasHeal(effect: CardEffect): boolean {
-  switch (effect.type) {
-    case EFFECT_GAIN_HEALING:
-      return true;
-
-    case EFFECT_CHOICE:
-      return effect.options.some(opt => effectHasHeal(opt));
-
-    case EFFECT_COMPOUND:
-      return effect.effects.some(eff => effectHasHeal(eff));
-
-    case EFFECT_CONDITIONAL:
-      return effectHasHeal(effect.thenEffect) ||
-        (effect.elseEffect ? effectHasHeal(effect.elseEffect) : false);
-
-    case EFFECT_SCALING:
-      return effectHasHeal(effect.baseEffect);
-
-    default:
-      return false;
-  }
-}
-
-/**
- * Check if an effect draws cards.
- */
-function effectHasDraw(effect: CardEffect): boolean {
-  switch (effect.type) {
-    case EFFECT_DRAW_CARDS:
-      return true;
-
-    case EFFECT_CHOICE:
-      return effect.options.some(opt => effectHasDraw(opt));
-
-    case EFFECT_COMPOUND:
-      return effect.effects.some(eff => effectHasDraw(eff));
-
-    case EFFECT_CONDITIONAL:
-      return effectHasDraw(effect.thenEffect) ||
-        (effect.elseEffect ? effectHasDraw(effect.elseEffect) : false);
-
-    case EFFECT_SCALING:
-      return effectHasDraw(effect.baseEffect);
-
-    default:
-      return false;
-  }
-}
-
-/**
- * Check if an effect applies a modifier (e.g., Mana Draw's extra source die).
- * These are "special" effects that don't fit the standard move/influence/heal/draw categories.
- */
-function effectHasModifier(effect: CardEffect): boolean {
-  switch (effect.type) {
-    case EFFECT_APPLY_MODIFIER:
-      return true;
-
-    case EFFECT_CHOICE:
-      return effect.options.some(opt => effectHasModifier(opt));
-
-    case EFFECT_COMPOUND:
-      return effect.effects.some(eff => effectHasModifier(eff));
-
-    case EFFECT_CONDITIONAL:
-      return effectHasModifier(effect.thenEffect) ||
-        (effect.elseEffect ? effectHasModifier(effect.elseEffect) : false);
-
-    case EFFECT_SCALING:
-      return effectHasModifier(effect.baseEffect);
-
-    default:
-      return false;
-  }
-}
-
-/**
- * Check if an effect gains mana tokens (e.g., Concentration).
- */
-function effectHasManaGain(effect: CardEffect): boolean {
-  switch (effect.type) {
-    case EFFECT_GAIN_MANA:
-      return true;
-
-    case EFFECT_CHOICE:
-      return effect.options.some(opt => effectHasManaGain(opt));
-
-    case EFFECT_COMPOUND:
-      return effect.effects.some(eff => effectHasManaGain(eff));
-
-    case EFFECT_CONDITIONAL:
-      return effectHasManaGain(effect.thenEffect) ||
-        (effect.elseEffect ? effectHasManaGain(effect.elseEffect) : false);
-
-    case EFFECT_SCALING:
-      return effectHasManaGain(effect.baseEffect);
-
-    default:
-      return false;
-  }
-}
-
-/**
- * Check if an effect is the Mana Draw powered effect (take dice, set colors, gain mana).
- * This special effect allows the player to manipulate source dice directly.
- */
-function effectHasManaDrawPowered(effect: CardEffect): boolean {
-  switch (effect.type) {
-    case EFFECT_MANA_DRAW_POWERED:
-      return true;
-
-    case EFFECT_CHOICE:
-      return effect.options.some(opt => effectHasManaDrawPowered(opt));
-
-    case EFFECT_COMPOUND:
-      return effect.effects.some(eff => effectHasManaDrawPowered(eff));
-
-    case EFFECT_CONDITIONAL:
-      return effectHasManaDrawPowered(effect.thenEffect) ||
-        (effect.elseEffect ? effectHasManaDrawPowered(effect.elseEffect) : false);
-
-    case EFFECT_SCALING:
-      return effectHasManaDrawPowered(effect.baseEffect);
-
-    default:
-      return false;
-  }
-}
-
-/**
- * Check if an effect involves crystal manipulation (gain crystal or convert mana to crystal).
- * These are "special" effects from Crystallize and similar cards.
- */
-function effectHasCrystal(effect: CardEffect): boolean {
-  switch (effect.type) {
-    case EFFECT_GAIN_CRYSTAL:
-    case EFFECT_CONVERT_MANA_TO_CRYSTAL:
-      return true;
-
-    case EFFECT_CHOICE:
-      return effect.options.some(opt => effectHasCrystal(opt));
-
-    case EFFECT_COMPOUND:
-      return effect.effects.some(eff => effectHasCrystal(eff));
-
-    case EFFECT_CONDITIONAL:
-      return effectHasCrystal(effect.thenEffect) ||
-        (effect.elseEffect ? effectHasCrystal(effect.elseEffect) : false);
-
-    case EFFECT_SCALING:
-      return effectHasCrystal(effect.baseEffect);
-
-    default:
-      return false;
-  }
-}
-
-/**
- * Check if an effect is a card boost effect (Concentration, Will Focus).
- * These effects let you play another Action card with its powered effect for free.
- */
-function effectHasCardBoost(effect: CardEffect): boolean {
-  switch (effect.type) {
-    case EFFECT_CARD_BOOST:
-      return true;
-
-    case EFFECT_CHOICE:
-      return effect.options.some(opt => effectHasCardBoost(opt));
-
-    case EFFECT_COMPOUND:
-      return effect.effects.some(eff => effectHasCardBoost(eff));
-
-    case EFFECT_CONDITIONAL:
-      return effectHasCardBoost(effect.thenEffect) ||
-        (effect.elseEffect ? effectHasCardBoost(effect.elseEffect) : false);
-
-    case EFFECT_SCALING:
-      return effectHasCardBoost(effect.baseEffect);
-
-    default:
-      return false;
-  }
-}
-
-/**
- * Check if an effect has enemy targeting (e.g., Tremor, Chill, Whirlwind).
- * These effects target specific enemies during combat to apply modifiers or defeat them.
- */
-function effectHasEnemyTargeting(effect: CardEffect): boolean {
-  switch (effect.type) {
-    case EFFECT_SELECT_COMBAT_ENEMY:
-      return true;
-
-    case EFFECT_CHOICE:
-      return effect.options.some(opt => effectHasEnemyTargeting(opt));
-
-    case EFFECT_COMPOUND:
-      return effect.effects.some(eff => effectHasEnemyTargeting(eff));
-
-    case EFFECT_CONDITIONAL:
-      return effectHasEnemyTargeting(effect.thenEffect) ||
-        (effect.elseEffect ? effectHasEnemyTargeting(effect.elseEffect) : false);
-
-    case EFFECT_SCALING:
-      return effectHasEnemyTargeting(effect.baseEffect);
-
-    default:
-      return false;
-  }
-}
