+++ packages/core/src/engine/validActions/cards/combat.ts
@@ -1,0 +1,156 @@
+/**
+ * Card playability computation for combat.
+ *
+ * Determines which cards in the player's hand can be played during each combat phase.
+ */
+
+import type { GameState } from "../../../state/GameState.js";
+import type { Player } from "../../../types/player.js";
+import type { CombatState, CombatPhase } from "../../../types/combat.js";
+import type { DeedCard } from "../../../types/cards.js";
+import type { PlayCardOptions, PlayableCard, ManaColor, SidewaysOption } from "@mage-knight/shared";
+import {
+  PLAY_SIDEWAYS_AS_ATTACK,
+  PLAY_SIDEWAYS_AS_BLOCK,
+} from "@mage-knight/shared";
+import {
+  COMBAT_PHASE_RANGED_SIEGE,
+  COMBAT_PHASE_BLOCK,
+  COMBAT_PHASE_ATTACK,
+} from "../../../types/combat.js";
+import { DEED_CARD_TYPE_WOUND, DEED_CARD_TYPE_SPELL } from "../../../types/cards.js";
+import { describeEffect } from "../../effects/describeEffect.js";
+import { isEffectResolvable } from "../../effects/index.js";
+import { getCard } from "./index.js";
+import { canPayForSpellBasic, findPayableManaColor } from "./manaPayment.js";
+import {
+  effectHasRangedOrSiege,
+  effectHasBlock,
+  effectHasAttack,
+  effectIsUtility,
+} from "./effectDetection/index.js";
+
+interface CardPlayability {
+  canPlayBasic: boolean;
+  canPlayPowered: boolean;
+  canPlaySideways: boolean;
+  sidewaysOptions: SidewaysOption[];
+}
+
+/**
+ * Get playable cards for combat based on the current phase.
+ */
+export function getPlayableCardsForCombat(
+  state: GameState,
+  player: Player,
+  combat: CombatState
+): PlayCardOptions {
+  const cards: PlayableCard[] = [];
+
+  for (const cardId of player.hand) {
+    const card = getCard(cardId);
+    if (!card) continue;
+
+    // Wounds cannot be played
+    if (card.cardType === DEED_CARD_TYPE_WOUND) continue;
+
+    const playability = getCardPlayabilityForPhase(card, combat.phase);
+
+    // Check resolvability - effect must actually be able to do something
+    const basicIsResolvable = isEffectResolvable(state, player.id, card.basicEffect);
+    const poweredIsResolvable = isEffectResolvable(state, player.id, card.poweredEffect);
+
+    // For spells, basic effect also requires mana (the spell's color)
+    // Get the spell's color from poweredBy (excluding black)
+    const spellBasicManaAvailable = card.cardType === DEED_CARD_TYPE_SPELL
+      ? canPayForSpellBasic(state, player, card)
+      : true; // Action cards don't need mana for basic effect
+
+    // Can only play basic if the phase allows it AND the effect is resolvable
+    // AND for spells, the player has the spell's color mana
+    const canActuallyPlayBasic = playability.canPlayBasic && basicIsResolvable && spellBasicManaAvailable;
+
+    // Check if the card has a powered effect for this phase AND player can pay for it AND it's resolvable
+    const payableManaColor = (playability.canPlayPowered && poweredIsResolvable)
+      ? findPayableManaColor(state, player, card)
+      : undefined;
+    const canActuallyPlayPowered = payableManaColor !== undefined;
+
+    if (canActuallyPlayBasic || canActuallyPlayPowered || playability.canPlaySideways) {
+      const playableCard: PlayableCard = {
+        cardId,
+        name: card.name,
+        canPlayBasic: canActuallyPlayBasic,
+        canPlayPowered: canActuallyPlayPowered,
+        canPlaySideways: playability.canPlaySideways,
+        basicEffectDescription: describeEffect(card.basicEffect),
+        poweredEffectDescription: describeEffect(card.poweredEffect),
+      };
+
+      // Only add optional properties when they have values
+      if (payableManaColor && canActuallyPlayPowered) {
+        (playableCard as { requiredMana?: ManaColor }).requiredMana = payableManaColor;
+      }
+      if (card.cardType === DEED_CARD_TYPE_SPELL) {
+        (playableCard as { isSpell?: boolean }).isSpell = true;
+      }
+      if (playability.sidewaysOptions && playability.sidewaysOptions.length > 0) {
+        (playableCard as { sidewaysOptions?: readonly SidewaysOption[] }).sidewaysOptions = playability.sidewaysOptions;
+      }
+
+      cards.push(playableCard);
+    }
+  }
+
+  return { cards };
+}
+
+/**
+ * Determine if a card can be played in a specific combat phase.
+ */
+function getCardPlayabilityForPhase(
+  card: DeedCard,
+  phase: CombatPhase
+): CardPlayability {
+  switch (phase) {
+    case COMBAT_PHASE_RANGED_SIEGE:
+      return {
+        // Ranged/siege phase: can play for ranged/siege attack OR utility effects
+        canPlayBasic: effectHasRangedOrSiege(card.basicEffect) || effectIsUtility(card.basicEffect),
+        canPlayPowered: effectHasRangedOrSiege(card.poweredEffect) || effectIsUtility(card.poweredEffect),
+        canPlaySideways: false, // Can't play sideways for ranged/siege
+        sidewaysOptions: [],
+      };
+
+    case COMBAT_PHASE_BLOCK:
+      return {
+        // Block phase: can play for block OR utility effects
+        canPlayBasic: effectHasBlock(card.basicEffect) || effectIsUtility(card.basicEffect),
+        canPlayPowered: effectHasBlock(card.poweredEffect) || effectIsUtility(card.poweredEffect),
+        canPlaySideways: card.sidewaysValue > 0,
+        sidewaysOptions: card.sidewaysValue > 0
+          ? [{ as: PLAY_SIDEWAYS_AS_BLOCK, value: card.sidewaysValue }]
+          : [],
+      };
+
+    case COMBAT_PHASE_ATTACK:
+      return {
+        // Attack phase: can play for attack OR utility effects
+        canPlayBasic: effectHasAttack(card.basicEffect) || effectIsUtility(card.basicEffect),
+        canPlayPowered: effectHasAttack(card.poweredEffect) || effectIsUtility(card.poweredEffect),
+        canPlaySideways: card.sidewaysValue > 0,
+        sidewaysOptions: card.sidewaysValue > 0
+          ? [{ as: PLAY_SIDEWAYS_AS_ATTACK, value: card.sidewaysValue }]
+          : [],
+      };
+
+    default:
+      // ASSIGN_DAMAGE phase - no cards played
+      return {
+        canPlayBasic: false,
+        canPlayPowered: false,
+        canPlaySideways: false,
+        sidewaysOptions: [],
+      };
+  }
+}
