+++ packages/core/src/engine/validActions/cards/effectDetection/combatEffects.ts
@@ -1,0 +1,97 @@
+/**
+ * Combat effect detection functions.
+ *
+ * Functions to detect attack, block, ranged, and siege effects in cards.
+ */
+
+import type { CardEffect } from "../../../../types/cards.js";
+import {
+  EFFECT_GAIN_ATTACK,
+  EFFECT_GAIN_BLOCK,
+  EFFECT_CHOICE,
+  EFFECT_COMPOUND,
+  EFFECT_CONDITIONAL,
+  EFFECT_SCALING,
+} from "../../../../types/effectTypes.js";
+import {
+  COMBAT_TYPE_RANGED,
+  COMBAT_TYPE_SIEGE,
+} from "@mage-knight/shared";
+
+/**
+ * Check if an effect provides ranged or siege attack.
+ */
+export function effectHasRangedOrSiege(effect: CardEffect): boolean {
+  switch (effect.type) {
+    case EFFECT_GAIN_ATTACK:
+      return effect.combatType === COMBAT_TYPE_RANGED || effect.combatType === COMBAT_TYPE_SIEGE;
+
+    case EFFECT_CHOICE:
+      return effect.options.some(opt => effectHasRangedOrSiege(opt));
+
+    case EFFECT_COMPOUND:
+      return effect.effects.some(eff => effectHasRangedOrSiege(eff));
+
+    case EFFECT_CONDITIONAL:
+      return effectHasRangedOrSiege(effect.thenEffect) ||
+        (effect.elseEffect ? effectHasRangedOrSiege(effect.elseEffect) : false);
+
+    case EFFECT_SCALING:
+      return effectHasRangedOrSiege(effect.baseEffect);
+
+    default:
+      return false;
+  }
+}
+
+/**
+ * Check if an effect provides block.
+ */
+export function effectHasBlock(effect: CardEffect): boolean {
+  switch (effect.type) {
+    case EFFECT_GAIN_BLOCK:
+      return true;
+
+    case EFFECT_CHOICE:
+      return effect.options.some(opt => effectHasBlock(opt));
+
+    case EFFECT_COMPOUND:
+      return effect.effects.some(eff => effectHasBlock(eff));
+
+    case EFFECT_CONDITIONAL:
+      return effectHasBlock(effect.thenEffect) ||
+        (effect.elseEffect ? effectHasBlock(effect.elseEffect) : false);
+
+    case EFFECT_SCALING:
+      return effectHasBlock(effect.baseEffect);
+
+    default:
+      return false;
+  }
+}
+
+/**
+ * Check if an effect provides any attack (melee, ranged, or siege).
+ */
+export function effectHasAttack(effect: CardEffect): boolean {
+  switch (effect.type) {
+    case EFFECT_GAIN_ATTACK:
+      return true;
+
+    case EFFECT_CHOICE:
+      return effect.options.some(opt => effectHasAttack(opt));
+
+    case EFFECT_COMPOUND:
+      return effect.effects.some(eff => effectHasAttack(eff));
+
+    case EFFECT_CONDITIONAL:
+      return effectHasAttack(effect.thenEffect) ||
+        (effect.elseEffect ? effectHasAttack(effect.elseEffect) : false);
+
+    case EFFECT_SCALING:
+      return effectHasAttack(effect.baseEffect);
+
+    default:
+      return false;
+  }
+}
